+++ .vscode/settings.json
@@ -1,0 +1,12 @@
+{
+    "java.project.sourcePaths": [
+        "src",
+        "src/main/java",
+        "src/test/java"
+    ],
+    "java.project.referencedLibraries": [],
+    "java.format.settings.url": ".vscode/java-formatter.xml",
+    "[java]": {
+        "editor.tabSize": 3
+    }
+}
