+++ build.gradle
@@ -1,7 +1,3 @@
-// https://github.com/junit-team/junit5-samples/blob/r5.11.0/junit5-jupiter-starter-gradle/build.gradle
-// https://symflower.com/en/company/blog/2023/running-junit-tests-with-gradle/
-// https://docs.gradle.org/current/dsl/org.gradle.api.tasks.testing.Test.html
-
 plugins {
 	id 'java'
 	// id 'eclipse' // optional (to generate Eclipse project files)
