+++ src/main/java/Blowfish/UInt128.java
@@ -1,3 +1,24 @@
+/**
+ *   UInt128.java
+ *
+ *   A monad-like abstraction for bitwise integer 128 bits operations.
+ *
+ *   Copyright (C) Aug, 2024 - Olaf Reitmaier Veracierta <olafrv@gmail.com>
+ *
+ *   This program is free software: you can redistribute it and/or modify
+ *   it under the terms of the GNU General Public License as published by
+ *   the Free Software Foundation, either version 3 of the License, or
+ *   (at your option) any later version.
+ *
+ *   This program is distributed in the hope that it will be useful,
+ *   but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *   GNU General Public License for more details.
+ *
+ *   You should have received a copy of the GNU General Public License
+ *   along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
 package Blowfish;
 
 public class UInt128 {
@@ -46,7 +67,7 @@ public class UInt128 {
         return new UInt128(0, high >>> (n - 64));
     }
 
-    /*
+    /* Prototype, untested!
     public UInt128 add(UInt128 other) {
         long newLow = this.low + other.low;
         long carry = (newLow < this.low) ? 1 : 0;
